Remove old table. (#392)
