--- GraphsV2/NewTable/default.config.js
@@ -1,13 +1,0 @@
-
-export const properties = {
-    width: "100%",
-    padding: 5,
-    limit: 100,
-    fixedHeader: true,
-    selectable: true,
-    multiSelectable: false,
-    showCheckboxes: false,
-    enableSelectAll: true,
-    headerHeight: 50,
-    rowHeight: 40,
-}
