--- GraphsV2/NewTable/style.js
@@ -1,14 +1,0 @@
-const tableStyle = {
-    design: {
-        margin: "10px 5px 0px 0px",
-        float: 'right',
-        maxHeight: '0px',
-        maxWidth: '30px',
-        padding: '0',
-        top: "40px",
-        right: '10px',
-        position: "fixed"
-    }
-}
-
-export default tableStyle;
