+++ GraphsV2/Table/default.config.js
@@ -1,4 +1,3 @@
-import { theme } from "../../theme";
 
 export const properties = {
     width: "100%",
@@ -9,26 +8,6 @@ export const properties = {
     multiSelectable: false,
     showCheckboxes: false,
     enableSelectAll: true,
-    headerPadding: 10,
+    headerHeight: 50,
     rowHeight: 40,
-    fontColor: theme.palette.blackLightColor,
-    highlightColor: theme.palette.greenLighterColor,
-    border: {
-        top: "solid 1px " + theme.palette.greyLightColor,
-        left: "0",
-        right: "0",
-        bottom: "0",
-    },
-    header: {
-        fontColor: theme.palette.blackColor,
-        border: {
-            top: "0",
-            left: "0",
-            right: "0",
-            bottom: "solid 2px " + theme.palette.greyLightColor,
-        }
-    },
-    tableStyle: {
-        tableLayout: "auto"
-    }
 }
