--- GraphsV2/Table/mock/configurations/hidePagination.json
@@ -1,12 +1,0 @@
-{
-    "data": {
-        "hidePagination": true,
-        "searchBar": false,
-        "columns": [
-            { "column": "Timestamp", "label": "Timestamp", "timeFormat": "%b %d, %y %X"},
-            { "column": "Application", "label": "Application" },
-            { "column": "MEMORY", "label": "MEMORY" },
-            { "column": "Value", "label": "Value" }
-        ]
-    }
-}
