--- GraphsV2/Table/mock/configurations/pagination.json
@@ -1,20 +1,0 @@
-{
-    "data": {
-        "id" : "pagination",
-        "searchBar": false,
-        "columns": [
-            { "column": "Timestamp", "label": "Timestamp", "timeFormat": "%b %d, %y %X"},
-            { "column": "Application", "label": "Application" },
-            { "column": "MEMORY", "label": "MEMORY" },
-            { "column": "Value", "label": "Value" }
-        ]
-    },
-    "listeners": [
-        {
-            "redirect": "/dashboards/medalsPerCountry",
-            "params": {
-                "app": "value"
-            }
-        }
-    ]
-}
