--- GraphsV2/Table/mock/configurations/searchWithOrOperator.json
@@ -1,11 +1,0 @@
-{
-    "data": {
-        "searchText": "Application == Developing OR Application == Testing",
-        "columns": [
-            { "column": "Timestamp", "label": "Timestamp", "timeFormat": "%b %d, %y %X"},
-            { "column": "Application", "label": "Application" },
-            { "column": "MEMORY", "label": "MEMORY" },
-            { "column": "Value", "label": "Value" }
-        ]
-    }
-}
