--- GraphsV2/Table/mock/configurations/selectColumn.json
@@ -1,26 +1,0 @@
-{
-    "data": {
-        "searchBar": false,
-        "selectColumnOption": true,
-        "multiSelectable": true,
-        "columns": [
-            {
-                "column": "Timestamp",
-                "label": "Timestamp",
-                "timeFormat": "%b %d, %y %X"
-            },
-            {
-                "column": "Application",
-                "label": "Application"
-            },
-            {
-                "column": "MEMORY",
-                "label": "MEMORY"
-            },
-            {
-                "column": "Value",
-                "label": "Value"
-            }
-        ]
-    }
-}
