--- GraphsV2/Table/mock/configurations/showInfoBox.json
@@ -1,12 +1,0 @@
-{
-    "data": {
-        "searchText": "Application == Developing",
-        "showInfoBox": true,
-        "columns": [
-            { "column": "Timestamp", "label": "Timestamp", "timeFormat": "%b %d, %y %X"},
-            { "column": "Application", "label": "Application", "infoBox": true },
-            { "column": "MEMORY", "label": "MEMORY", "infoBox": true },
-            { "column": "Value", "label": "Value", "infoBox": true }
-        ]
-    }
-}
