--- GraphsV2/Table/style.css
@@ -1,40 +1,0 @@
-.pointer {
-    cursor: pointer !important;
-    text-decoration: none;
-}
-
-.text-white {
-    color: #fcfcfc;
-}
-
-.text-white:hover {
-    font-weight: bold;
-    color: #fcfcfc;
-    cursor:  copy;
-}
-
-table tbody tr:hover {
-    background-color: #f2f2f2;
-    cursor: pointer;
-}
-
-table th > div:nth-child(2) {
-    margin-left: 1px !important;
-    margin-right: 0px !important;
-    height: 10px !important;
-    width: 10px !important;
-}
-
-.dialogBody > div {
-    box-shadow: rgba(0, 0, 0, 0.25) 0px 14px 45px, rgba(0, 0, 0, 0.22) 0px 10px 18px !important;
-}
-.wrapper-data{
-    white-space: nowrap;
-    position: relative;
-}
-.wrapper-data > div:hover{
-    z-index: 99999 !important;
-}
-.refreshIcon{
-    height: 20px !important;
-}
